+++ Ktisis/Services/InteropService.cs
@@ -1,15 +1,21 @@
-using System;
+﻿using System;
 using System.Numerics;
+using System.Collections.Generic;
 using System.Runtime.InteropServices;
+
+using Dalamud.Hooking;
+using Dalamud.Logging;
 
 using FFXIVClientStructs.Havok;
 
-namespace Ktisis.Interop {
-	internal static class Alloc {
+namespace Ktisis.Services {
+	internal static class InteropService {
 		// Allocations
+
 		private static IntPtr MatrixAlloc;
 
-		// Access
+		// Matrix Access
+
 		internal unsafe static Matrix4x4* Matrix; // Align to 16-byte boundary
 		internal unsafe static Matrix4x4 GetMatrix(hkQsTransformf* transform) {
 			transform->get4x4ColumnMajor((float*)Matrix);
@@ -20,15 +26,24 @@ namespace Ktisis.Interop {
 			transform->set((hkMatrix4f*)Matrix);
 		}
 
-		// Init & disspose
-		public unsafe static void Init() {
+		// Hooks
+
+		private static List<Hook<Delegate>> Hooks = new();
+
+		internal unsafe static void Init() {
 			// Allocate space for our matrix to be aligned on a 16-byte boundary.
 			// This is required due to ffxiv's use of the MOVAPS instruction.
 			// Thanks to Fayti1703 for helping with debugging and coming up with this fix.
 			MatrixAlloc = Marshal.AllocHGlobal(sizeof(float) * 16 + 16);
 			Matrix = (Matrix4x4*)(16 * ((long)(MatrixAlloc + 15) / 16));
+
+			PluginLog.Debug(string.Format(
+				"Allocated matrix at {0:X}, aligned at {1:X}.",
+				MatrixAlloc, (nint)Matrix
+			));
 		}
-		public static void Dispose() {
+
+		internal static void Dispose() {
 			Marshal.FreeHGlobal(MatrixAlloc);
 		}
 	}
